add API changes to Resource Service and Test - refactor resource creation and update type
